indexesview/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1
djangorestframework==3.13.1
drf-yasg==1.21.3
filelock==3.7.1
frozenlist==1.3.1
future==0.18.2
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.9.1
kombu==5.2.4
MarkupSafe==2.1.1
multidict==6.0.2
openapi-codec==1.3.2
packaging==21.3
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
pycoingecko==2.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
uritools==4.0.0
urlextract==1.6.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
